introduced another layer of abstraction AbstractUnitOfWork to AbstractTemplatesUnitOfWork in order to remove unnecessary code. Added __init__.py for said abstraction and added an import of it to AbstractTemplatesUnitOfWork
