+++ backend/modules/common/ports/unit_of_work/__init__.py
@@ -1,0 +1,3 @@
+from .unit_of_work import AbstractUnitOfWork
+
+__all__ = ["AbstractUnitOfWork"]
