+++ backend/modules/common/ports/unit_of_work/unit_of_work.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+import abc
+
+class AbstractUnitOfWork(abc.ABC):
+    def __enter__(self) -> AbstractUnitOfWork:
+        return self
+
+    def __exit__(self, *args):
+        self.rollback()
+
+    @abc.abstractmethod
+    def commit(self):
+        pass
+
+    @abc.abstractmethod
+    def rollback(self):
+        pass
